refactor: Rename ConfigurationManager to legacy and replace with V2 version
