+++ src/test/java/com/synopsys/integration/alert/performance/ScalingExternalPerformanceTest.java
@@ -21,7 +21,7 @@ import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
 import com.synopsys.integration.alert.descriptor.api.JiraServerChannelKey;
 import com.synopsys.integration.alert.performance.model.PerformanceExecutionStatusModel;
 import com.synopsys.integration.alert.performance.utility.BlackDuckProviderService;
-import com.synopsys.integration.alert.performance.utility.ConfigurationManagerV2;
+import com.synopsys.integration.alert.performance.utility.ConfigurationManager;
 import com.synopsys.integration.alert.performance.utility.ExternalAlertRequestUtility;
 import com.synopsys.integration.alert.performance.utility.IntegrationPerformanceTestRunner;
 import com.synopsys.integration.alert.performance.utility.PerformanceLoggingUtility;
@@ -54,7 +54,7 @@ class ScalingExternalPerformanceTest {
     private BlackDuckProviderService blackDuckProviderService;
     private JiraServerPerformanceUtility jiraServerPerformanceUtility;
     private IntegrationPerformanceTestRunner testRunner;
-    private ConfigurationManagerV2 configurationManager;
+    private ConfigurationManager configurationManager;
 
     private int numberOfProjectsToCreate;
     private int numberOfJobsToCreate;
@@ -74,7 +74,7 @@ class ScalingExternalPerformanceTest {
         ExternalAlertRequestUtility alertRequestUtility = new ExternalAlertRequestUtility(intLogger, client, alertURL);
         alertRequestUtility.loginToExternalAlert();
         blackDuckProviderService = new BlackDuckProviderService(alertRequestUtility, gson);
-        configurationManager = new ConfigurationManagerV2(
+        configurationManager = new ConfigurationManager(
             gson,
             alertRequestUtility,
             blackDuckProviderService.getBlackDuckProviderKey(),
