+++ src/test/java/com/synopsys/integration/alert/performance/utility/ConfigurationManager.java
@@ -3,13 +3,16 @@ package com.synopsys.integration.alert.performance.utility;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.slf4j.LoggerFactory;
+
 import com.google.gson.Gson;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.synopsys.integration.alert.api.common.model.Obfuscated;
+import com.synopsys.integration.alert.api.common.model.ValidationResponseModel;
 import com.synopsys.integration.alert.api.common.model.exception.AlertRuntimeException;
 import com.synopsys.integration.alert.api.provider.ProviderDescriptor;
 import com.synopsys.integration.alert.common.descriptor.ChannelDescriptor;
@@ -20,11 +23,15 @@ import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
 import com.synopsys.integration.alert.common.rest.model.JobFieldModel;
 import com.synopsys.integration.alert.common.rest.model.JobPagedModel;
 import com.synopsys.integration.alert.common.rest.model.JobProviderProjectFieldModel;
-import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
 import com.synopsys.integration.alert.descriptor.api.model.ChannelKeys;
+import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
+import com.synopsys.integration.blackduck.service.model.ProjectVersionWrapper;
 import com.synopsys.integration.exception.IntegrationException;
+import com.synopsys.integration.log.IntLogger;
+import com.synopsys.integration.log.Slf4jIntLogger;
 
 public class ConfigurationManager {
+    private final IntLogger intLogger = new Slf4jIntLogger(LoggerFactory.getLogger(getClass()));
     private final AlertRequestUtility alertRequestUtility;
     private final Gson gson;
     private final String blackDuckProviderKey;
@@ -37,51 +44,126 @@ public class ConfigurationManager {
         this.channelKey = channelKey;
     }
 
-    public void createGlobalConfiguration(FieldModel globalConfig) {
-        String configurationRequestBody = gson.toJson(globalConfig);
-
-        String descriptorName = globalConfig.getDescriptorName();
+    public <T extends Obfuscated<T>> Optional<T> createGlobalConfiguration(String apiConfigurationPath, T globalConfigModel, Class<T> modelType) {
         try {
-            String globalConfigSearchResponse = alertRequestUtility.executeGetRequest(
-                String.format("/api/configuration?context=%s&descriptorName=%s", ConfigContextEnum.GLOBAL.name(), descriptorName),
-                String.format("Could not find the existing global configuration for %s.", descriptorName));
-            JsonObject globalConfigSearchJsonObject = gson.fromJson(globalConfigSearchResponse, JsonObject.class);
-            JsonArray fieldModels = globalConfigSearchJsonObject.get("fieldModels").getAsJsonArray();
-            JsonElement firstFieldModel = fieldModels.get(0);
-            FieldModel originalGlobalConfig = gson.fromJson(firstFieldModel, FieldModel.class);
+            String requestBody = gson.toJson(globalConfigModel);
 
-            globalConfig.setId(originalGlobalConfig.getId());
+            String validationResponseString = alertRequestUtility
+                .executePostRequest(String.format("%s/validate", apiConfigurationPath), requestBody, "Validating the global configuration failed.");
+            ValidationResponseModel validationResponse = gson.fromJson(validationResponseString, ValidationResponseModel.class);
+            if (validationResponse.hasErrors()) {
+                intLogger.error(String.format("Could not validate global configuration model. Error: %s", validationResponse.getErrors()));
+                return Optional.empty();
+            }
+            String testResponseString = alertRequestUtility
+                .executePostRequest(String.format("%s/test", apiConfigurationPath), requestBody, "Testing the global configuration failed.");
+            ValidationResponseModel testResponse = gson.fromJson(testResponseString, ValidationResponseModel.class);
+            if (testResponse.hasErrors() && !ValidationResponseModel.VALIDATION_SUCCESS_MESSAGE.equals(validationResponse.getMessage())) {
+                intLogger.error(String.format("Testing the global config model error message: %s", validationResponse.getMessage()));
+                intLogger.error(String.format("Testing the global config model failed. Error: %s", validationResponse.getErrors()));
+                return Optional.empty();
+            }
 
-            alertRequestUtility.executePostRequest("/api/configuration/validate", configurationRequestBody, String.format("Validating the global configuration %s failed.", descriptorName));
-            alertRequestUtility.executePostRequest("/api/configuration/test", configurationRequestBody, String.format("Testing the global configuration %s failed.", descriptorName));
-            alertRequestUtility.executePutRequest(
-                String.format("/api/configuration/%s", globalConfig.getId()), configurationRequestBody,
-                String.format("Could not create the global configuration %s.", descriptorName));
+            String globalConfigCreateResponse = alertRequestUtility.executePostRequest(apiConfigurationPath, requestBody, "Could not create the global configuration");
+            JsonObject globalConfigSearchJsonObject = gson.fromJson(globalConfigCreateResponse, JsonObject.class);
+            T savedGlobalConfig = gson.fromJson(globalConfigSearchJsonObject, modelType);
+            return Optional.of(savedGlobalConfig);
         } catch (IntegrationException e) {
-            throw new RuntimeException(e.getMessage(), e);
+            intLogger.error("Unexpected error occurred while creating the global configuration.", e);
+            return Optional.empty();
         }
     }
 
     public String createJob(Map<String, FieldValueModel> channelFields, String jobName, String blackDuckProviderId, String blackDuckProjectName) throws IntegrationException {
-        return createJob(channelFields, jobName, blackDuckProviderId, blackDuckProjectName,
-            List.of(NotificationType.BOM_EDIT, NotificationType.POLICY_OVERRIDE, NotificationType.RULE_VIOLATION, NotificationType.RULE_VIOLATION_CLEARED, NotificationType.VULNERABILITY));
+        JobProviderProjectFieldModel providerProjectModel = new JobProviderProjectFieldModel(blackDuckProjectName, "href", false);
+        return createJob(channelFields, jobName, blackDuckProviderId, List.of(providerProjectModel),
+            List.of(
+                NotificationType.BOM_EDIT,
+                NotificationType.POLICY_OVERRIDE,
+                NotificationType.RULE_VIOLATION,
+                NotificationType.RULE_VIOLATION_CLEARED,
+                NotificationType.VULNERABILITY
+            )
+        );
     }
 
-    public String createJob(Map<String, FieldValueModel> channelFields, String jobName, String blackDuckProviderId, String blackDuckProjectName, List<NotificationType> notificationTypes) throws IntegrationException {
+    public String createJob(Map<String, FieldValueModel> channelFields, String jobName, String blackDuckProviderId, List<ProjectVersionWrapper> projectVersionWrappers)
+        throws IntegrationException {
+        List<JobProviderProjectFieldModel> providerProjectModels = projectVersionWrappers
+            .stream()
+            .map(ProjectVersionWrapper::getProjectView)
+            .map(projectView -> new JobProviderProjectFieldModel(projectView.getName(), projectView.getHref().toString(), false))
+            .collect(Collectors.toList());
+        return createJob(channelFields, jobName, blackDuckProviderId, providerProjectModels,
+            List.of(
+                NotificationType.BOM_EDIT,
+                NotificationType.POLICY_OVERRIDE,
+                NotificationType.RULE_VIOLATION,
+                NotificationType.RULE_VIOLATION_CLEARED,
+                NotificationType.VULNERABILITY
+            )
+        );
+    }
+
+    public String createJob(
+        Map<String, FieldValueModel> channelFields,
+        String jobName,
+        String blackDuckProviderId,
+        List<JobProviderProjectFieldModel> providerProjectModel,
+        List<NotificationType> notificationTypes
+    ) throws IntegrationException {
         List<String> notificationTypeNames = notificationTypes.stream()
             .map(Enum::name)
+            .collect(Collectors.toList());
+        List<String> blackDuckProjectNames = providerProjectModel
+            .stream()
+            .map(JobProviderProjectFieldModel::getName)
             .collect(Collectors.toList());
         Map<String, FieldValueModel> providerKeyToValues = new HashMap<>();
         providerKeyToValues.put(ProviderDescriptor.KEY_PROVIDER_CONFIG_ID, new FieldValueModel(List.of(blackDuckProviderId), true));
         providerKeyToValues.put(ProviderDescriptor.KEY_NOTIFICATION_TYPES, new FieldValueModel(notificationTypeNames, true));
         providerKeyToValues.put(ProviderDescriptor.KEY_PROCESSING_TYPE, new FieldValueModel(List.of(ProcessingType.DEFAULT.name()), true));
         providerKeyToValues.put(ProviderDescriptor.KEY_FILTER_BY_PROJECT, new FieldValueModel(List.of("true"), true));
-        providerKeyToValues.put(ProviderDescriptor.KEY_CONFIGURED_PROJECT, new FieldValueModel(List.of(blackDuckProjectName), true));
+        providerKeyToValues.put(ProviderDescriptor.KEY_CONFIGURED_PROJECT, new FieldValueModel(blackDuckProjectNames, true));
         FieldModel jobProviderConfiguration = new FieldModel(blackDuckProviderKey, ConfigContextEnum.DISTRIBUTION.name(), providerKeyToValues);
 
         FieldModel jobConfiguration = new FieldModel(channelKey, ConfigContextEnum.DISTRIBUTION.name(), channelFields);
 
-        JobFieldModel jobFieldModel = new JobFieldModel(null, Set.of(jobConfiguration, jobProviderConfiguration), List.of(new JobProviderProjectFieldModel(blackDuckProjectName, "href", false)));
+        JobFieldModel jobFieldModel = new JobFieldModel(
+            null,
+            Set.of(jobConfiguration, jobProviderConfiguration),
+            providerProjectModel
+        );
+
+        String jobConfigBody = gson.toJson(jobFieldModel);
+
+        alertRequestUtility.executePostRequest("/api/configuration/job/validate", jobConfigBody, String.format("Validating the Job %s failed.", jobName));
+        alertRequestUtility.executePostRequest("/api/configuration/job/test", jobConfigBody, String.format("Testing the Job %s failed.", jobName));
+        String creationResponse = alertRequestUtility.executePostRequest("/api/configuration/job", jobConfigBody, String.format("Could not create the Job %s.", jobName));
+
+        JsonObject jsonObject = gson.fromJson(creationResponse, JsonObject.class);
+        return jsonObject.get("jobId").getAsString();
+    }
+
+    public String createPolicyViolationJob(
+        Map<String, FieldValueModel> channelFields,
+        String jobName,
+        String blackDuckProviderId
+    ) throws IntegrationException {
+        Map<String, FieldValueModel> providerKeyToValues = new HashMap<>();
+        providerKeyToValues.put(ProviderDescriptor.KEY_PROVIDER_CONFIG_ID, new FieldValueModel(List.of(blackDuckProviderId), true));
+        providerKeyToValues.put(ProviderDescriptor.KEY_NOTIFICATION_TYPES, new FieldValueModel(List.of(NotificationType.RULE_VIOLATION.name()), true));
+        providerKeyToValues.put(ProviderDescriptor.KEY_PROCESSING_TYPE, new FieldValueModel(List.of(ProcessingType.DEFAULT.name()), true));
+        providerKeyToValues.put(ProviderDescriptor.KEY_FILTER_BY_PROJECT, new FieldValueModel(List.of("false"), true));
+        FieldModel jobProviderConfiguration = new FieldModel(blackDuckProviderKey, ConfigContextEnum.DISTRIBUTION.name(), providerKeyToValues);
+
+        FieldModel jobConfiguration = new FieldModel(channelKey, ConfigContextEnum.DISTRIBUTION.name(), channelFields);
+
+        JobFieldModel jobFieldModel = new JobFieldModel(
+            null,
+            Set.of(jobConfiguration, jobProviderConfiguration),
+            List.of()
+        );
 
         String jobConfigBody = gson.toJson(jobFieldModel);
 
@@ -94,7 +176,8 @@ public class ConfigurationManager {
     }
 
     public void copyJob(String jobToCopy, String newJobName) throws IntegrationException {
-        String response = alertRequestUtility.executeGetRequest(String.format("/api/configuration/job?searchTerm=%s", jobToCopy), String.format("Could not copy the Job %s.", jobToCopy));
+        String response = alertRequestUtility
+            .executeGetRequest(String.format("/api/configuration/job?searchTerm=%s", jobToCopy), String.format("Could not copy the Job %s.", jobToCopy));
         JobPagedModel jobModel = gson.fromJson(response, JobPagedModel.class);
         JobFieldModel jobFieldModel = jobModel.getJobs().stream()
             .findFirst()
@@ -113,5 +196,5 @@ public class ConfigurationManager {
         String jobConfigBody = gson.toJson(jobFieldModel);
         alertRequestUtility.executePostRequest("/api/configuration/job", jobConfigBody, String.format("Could not create the Job %s.", newJobName));
     }
-
 }
+
